refactor: DI 테스트 DI 모듈로 이동
