--- app/src/test/java/woowacourse/shopping/di/FakeObj.kt
@@ -1,3 +1,0 @@
-package woowacourse.shopping.di
-
-object FakeObj
