--- app/src/test/java/woowacourse/shopping/di/FakeRepository.kt
@@ -1,3 +1,0 @@
-package woowacourse.shopping.di
-
-interface FakeRepository
