--- app/src/test/java/woowacourse/shopping/di/SecondDataSource.kt
@@ -1,3 +1,0 @@
-package woowacourse.shopping.di
-
-interface SecondDataSource
