+++ berdi/src/test/java/com/di/berdi/fake/DefaultFirstDataSource.kt
@@ -1,3 +1,3 @@
-package woowacourse.shopping.di
+package com.di.berdi.fake
 
 class DefaultFirstDataSource : FirstDataSource
