+++ berdi/src/test/java/com/di/berdi/fake/FakeObj.kt
@@ -1,0 +1,3 @@
+package com.di.berdi.fake
+
+object FakeObj
