+++ berdi/src/test/java/com/di/berdi/fake/InMemoryFakeRepository.kt
@@ -1,3 +1,3 @@
-package woowacourse.shopping.di
+package com.di.berdi.fake
 
 class InMemoryFakeRepository : FakeRepository
