+++ berdi/src/test/java/com/di/berdi/fake/SecondDataSource.kt
@@ -1,0 +1,3 @@
+package com.di.berdi.fake
+
+interface SecondDataSource
